% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{as_category}
\alias{as_category}
\title{Validate and coerce any object as a categorical variable.}
\usage{
as_category(x, levels = c(na.omit(unique(x))))
}
\arguments{
\item{x}{A vector object to be coerced to categorical.}

\item{levels}{an optional vector of the unique values (as character strings)
that x might have taken. The default is the unique set of values taken by
as.character(x), sorted into increasing order of x. Note that this set can be
specified as smaller than sort(unique(x)).}
}
\value{
A vector with class haven_labelled.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Converts a vector object to a categorical object, typically a column in a
data frame. The categories come from valid values present in the
object and are added to an associated data dictionary (when present).
}
\examples{
{

library(dplyr)

##### Example 1: use madshapR_example provided by the package
dataset <- 
  madshapR_example$`dataset_example` \%>\%
  mutate(prg_ever = as_category(prg_ever))
  
head(dataset$prg_ever)

###### Example 2: any data frame can be a dataset
cat_cyl <- as_category(mtcars[['cyl']])

head(cat_cyl)

}

}
\seealso{
\code{\link[haven:labelled]{haven::labelled()}}
}
