% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataset_functions.R
\name{as_dossier}
\alias{as_dossier}
\title{Validate and coerce an object to dossier format}
\usage{
as_dossier(object)
}
\arguments{
\item{object}{A potential dossier to be coerced.}
}
\value{
A list of data frame(s) with \code{Rmonize::class} 'dossier'.
}
\description{
Confirms that the input object is a valid dossier and returns it as a dossier
with the appropriate \code{madshapR::class} attribute. This function mainly helps
validate input within other functions of the package but could be used to
check if a dossier is valid.
}
\details{
A dossier is a named list containing at least one data frame or more,
each of them being datasets. The name of each tibble will be use as the
reference name of the dataset.
}
\examples{
{

# use madshapR_DEMO provided by the package
library(stringr)

###### Example 1: a dataset list is a dossier by definition.
dossier <- 
  as_dossier(madshapR_DEMO[str_detect(names(madshapR_DEMO),"dataset_TOKYO")])
   
###### Example 2: any list of data frame can be a dossier by 
# definition.
as_dossier(list(dataset_1 = iris, dataset_2 = mtcars))

}

}
\seealso{
For a better assessment, please use \code{\link[=dataset_evaluate]{dataset_evaluate()}}.
}
