% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-import-from-fabR.R
\docType{import}
\name{bookdown_template}
\alias{bookdown_template}
\title{Objects exported from other packages}
\keyword{exported}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fabR}{\code{\link[fabR]{bookdown_template}}}
}}

