% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-unit_checks.R
\name{check_data_dict_categories}
\alias{check_data_dict_categories}
\title{Return categorical variables which has issues in a data dictionary}
\usage{
check_data_dict_categories(data_dict)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be evaluated.}
}
\value{
A tibble providing categorical variables that has issues within a
data dictionary.
}
\description{
Generates a tibble that reports any categorical variable which
name is in the 'Categories' but not present in the main list of variable
names in 'Variables' component. It also reports any categorical variable
which possible observation (the combination of 'variable' and 'name' in
'Categories') is not unique.
This tibble can be used to assist the user in the assessment of the data
structure, fields investigation (mandatory or not), coherence across elements
and taxonomy or standard evaluation.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
{

# use DEMO_files provided by the package

data_dict <- DEMO_files$`dd_TOKYO_format_maelstrom_tagged - ERROR`
check_data_dict_categories(data_dict)

}

}
