% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-unit_checks.R
\name{check_data_dict_taxonomy}
\alias{check_data_dict_taxonomy}
\title{Return evaluation of the valueType declared in a data dictionary}
\usage{
check_data_dict_taxonomy(data_dict, taxonomy)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be evaluated.}

\item{taxonomy}{A data frame or data frame extension (e.g. a tibble),
identifying the scheme used for variables classification as a tibble.}
}
\value{
A tibble providing non-standard taxonomy declared in a data dictionary.
}
\description{
Generates a tibble that reports any variable whose taxonomy is
not present in the list of authorized taxonomy list. This tibble can be used
to assist the user in the assessment of the data structure, fields
investigation (mandatory or not), coherence across elements and taxonomy or
standard evaluation.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A taxonomy must be a data frame or data frame extension (e.g. a tibble).
The taxonomy must be compatible with (and generally extracted from) an
Opal environment, and must contain at least 'taxonomy', 'vocabulary' and
'terms' to work with some specific functions. In addition, the taxonomy
may follow Maelstrom research stardards, and its content can be evaluated
accordingly, such as naming convention restriction, tagging elements,
or scales, which are specific to Maelstrom Research. In this particular
case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale',
'vocabulary_scale' and 'term_scale' to work with some specific functions.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.

data_dict <- data_dict_list$`dd_TOKYO_format_maelstrom_tagged - ERROR`

}

}
