% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataset_functions.R
\name{col_id}
\alias{col_id}
\title{Return the id column names(s) of a dataset}
\usage{
col_id(dataset)
}
\arguments{
\item{dataset}{A tibble identifying the input dataset observations.}
}
\value{
Name(s) of indetifier column(s). NULL if not.
}
\description{
Tests if the input object is a valid dossier. This function mainly helps
validate input within other functions of the package but could be used to
check if a dossier is valid.
}
\details{
A dataset must be a data frame-like object and can be associated with a
data dictionary. If no data dictionary is provided, a minimum workable
data dictionary will be generated as needed by relevant functions.
An identifier \code{id} column for sorting can be specified by the user. If
specified, the \code{id} values must be non-missing and will be used in functions
that require it. If no identifier column is specified, indexing is handled
automatically by the function.
}
\examples{
{

# use DEMO_files provided by the package
col_id(iris)
col_id(as_dataset(iris,'Species'))

}

}
