% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{color_palette_maelstrom}
\alias{color_palette_maelstrom}
\title{Built-in data frame of colors used in the graphs and charts.}
\format{
\subsection{\code{data.frame}}{

A data frame with 51 rows and 2 columns:
\describe{
\item{values}{possible class value in a dataset.}
\item{color_palette}{associated color}
...
}
}
}
\usage{
color_palette_maelstrom
}
\description{
Provides a built-in data frame of the colors used in the graphs and charts.
}
\examples{
{

head(color_palette_maelstrom)

}
}
\keyword{datasets}
