% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{data_dict_add_labels_short}
\alias{data_dict_add_labels_short}
\title{Add Shortened Labels to Data Dictionary}
\usage{
data_dict_add_labels_short(
  data_dict,
  max_length_name_var = 10,
  max_length_total_var = 25,
  max_length_name_cat = 10,
  max_length_total_cat = 15
)
}
\arguments{
\item{data_dict}{A data dictionary, typically a list containing 'Variables' and
'Categories' data frames.}

\item{max_length_name_var}{An integer specifying the maximum length for variable names
(default is 10).}

\item{max_length_total_var}{An integer specifying the maximum total length for variable
names and labels combined (default is 25).}

\item{max_length_name_cat}{An integer specifying the maximum length for category names
(default is 10).}

\item{max_length_total_cat}{An integer specifying the maximum total length for category
names and labels combined (default is 15).}
}
\value{
A modified data dictionary with additional columns for shortened labels:
\itemize{
\item \code{madshapR::label_short_var}: Shortened variable labels.
\item \code{madshapR::label_short_cat}: Shortened category labels (if categories are present).
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function modifies a data dictionary by adding shortened labels for both
variables and categories. The shortened labels are created based on specified
maximum lengths for the variable and category names and labels.
The function first validates the input using \code{as_data_dict_shape} and extracts the
first variable and category labels using \code{first_label_get}. It then calculates the
lengths of names and labels, ensuring that they do not exceed the specified maximum
lengths. The function handles both variables and categories, creating short labels
while replacing any missing values with "Empty".
}
\details{
A data dictionary contains the list of variables in a dataset and metadata
about the variables and can be associated with a dataset. A data dictionary
object is a list of data frame(s) named 'Variables' (required) and
'Categories' (if any). To be usable in any function, the data frame
'Variables' must contain at least the \code{name} column, with all unique and
non-missing entries, and the data frame 'Categories' must contain at least
the \code{variable} and \code{name} columns, with unique combination of
\code{variable} and \code{name}.
}
\examples{
{

 # use madshapR_example provided by the package
 data_dict <- madshapR_example$`data_dict_example`
 data_dict_with_short_labels <- data_dict_add_labels_short(data_dict,10,25,10,16)
 
 attributes(data_dict_with_short_labels)

}


}
