% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{data_dict_extract}
\alias{data_dict_extract}
\title{Create a data dictionary in the Maelstrom Research format from a dataset}
\usage{
data_dict_extract(dataset, as_data_dict_mlstr = TRUE)
}
\arguments{
\item{dataset}{A tibble identifying the input dataset observations which contains
meta data as attributes.}

\item{as_data_dict_mlstr}{Whether the output data dictionary has a simple
data dictionary structure or not (meaning has a Maelstrom data dictionary
structure, compatible with Maelstrom ecosystem such as Opal environment).
FALSE by default.}
}
\value{
A list of tibble(s) identifying a data dictionary.
}
\description{
Creates a data dictionary in the Maelstrom Research formats (with 'Variables'
and Categories' in separate tibbles and standard columns in each) from any
dataset in tibble format. If the input dataset has no associated metadata, a
data dictionary with a minimal of information is created from the column
(variable) names to create the data dictionary structure required for
madshapR (all columns except variable names will be blank).
}
\details{
A dataset must be a data-frame or data-frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data dictionary
can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research standards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
{

# use DEMO_files provided by the package

###### Example 2: extract data dictionary from a labelled dataset.
dataset <- DEMO_files$dataset_MELBOURNE_1
data_dict <- as_data_dict_mlstr(DEMO_files$dd_MELBOURNE_1_format_maelstrom)
dataset <- data_dict_apply(dataset,data_dict)
data_dict_extract(dataset)

###### Example 2: extract data dictionary from any dataset (the 
# data dictionary will be created upon attributes of the dataset. Factors 
# will be considered as categorical variables)
data_dict_extract(iris)

}

}
