% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{data_dict_filter}
\alias{data_dict_filter}
\title{Subset data dictionary using column values}
\usage{
data_dict_filter(
  data_dict,
  filter_var = NULL,
  filter_cat = NULL,
  filter_all = NULL
)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be
transformed.}

\item{filter_var}{Expressions that is defined in the element 'Variables' in
the data dictionary.}

\item{filter_cat}{Expressions that is defined in the element 'Categories' in
the data dictionary.}

\item{filter_all}{Expressions that is defined both in the 'Categories' and
'Variables' in the data dictionary.}
}
\value{
A list of tibble(s) identifying a data dictionary - like structure.
}
\description{
Subsets either or both the 'Variables' and the 'Categories' tibbles of a data
dictionary. Rows are conserved if the value of one column satisfy a
condition. This function is a wrapper analogous to \code{\link[dplyr:filter]{dplyr::filter()}}.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[dplyr:filter]{dplyr::filter()}}
}
