% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{data_dict_flatten}
\alias{data_dict_flatten}
\title{Transform any component in a data dictionary in colum(s) in another component}
\usage{
data_dict_flatten(
  data_dict,
  from = "Categories",
  to = "Variables",
  name_prefix = "Categories::"
)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be
transformed. Automatically generated if not provided.}

\item{from}{Symbol identifying the name of the component (tibble) to take
column(s) from. Default is 'Categories'.}

\item{to}{Symbol identifying the name of the component (tibble) to create
column(s) to. Default is 'Variables'.}

\item{name_prefix}{Character string of the prefix of columns of interest.
This prefix will be used to select columns, and to rename them in the 'to'
component. Default is 'Categories::'.}
}
\value{
A list of tibble(s) identifying a data dictionary.
}
\description{
If the component 'to' exists, and contains any column 'xx' or 'yy', these
columns will be added to the component 'from' under the names 'to:xx'
and 'to:yy'. (unique names will be generated if necessary). Each element
of these column will gather all information to process the reverse operation.
Separator of each element is the following structure :
'name = xx1 ; name = xx2'.
This function is the reversed operation of \code{\link[=data_dict_expand]{data_dict_expand()}}
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[=data_dict_expand]{data_dict_expand()}}
}
