% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dictionaries_functions.R
\name{data_dict_group_by}
\alias{data_dict_group_by}
\title{Group listed data dictionaries by specified column names}
\usage{
data_dict_group_by(data_dict, col)
}
\arguments{
\item{data_dict}{A list of data frame(s) representing metadata to be
transformed.}

\item{col}{variable to group by.}
}
\value{
A list of data frame(s) identifying a workable data dictionary structure.
}
\description{
Groups the data dictionary element(s) by the groups defined by the query.
This function groups both the 'Variables' and 'Categories' elements (if
the group exists under the same definition in in both). This function is
analogous to running \code{\link[dplyr:group_by]{dplyr::group_by()}}. Each element is named using the
group values. \code{\link[=data_dict_ungroup]{data_dict_ungroup()}} reverses the effect.
}
\details{
A data dictionary contains the list of variables in a dataset and metadata
about the variables and can be associated with a dataset. A data dictionary
object is a list of data frame(s) named 'Variables' (required) and
'Categories' (if any). To be usable in any function, the data frame
'Variables' must contain at least the \code{name} column, with all unique and
non-missing entries, and the data frame 'Categories' must contain at least
the \code{variable} and \code{name} columns, with unique combination of
\code{variable} and \code{name}.
}
\seealso{
\code{\link[dplyr:group_by]{dplyr::group_by()}}, \code{\link[=data_dict_ungroup]{data_dict_ungroup()}}
}
