% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dictionaries_functions.R
\name{data_dict_list_nest}
\alias{data_dict_list_nest}
\title{Bind listed data dictionaries}
\usage{
data_dict_list_nest(data_dict_list, name_group = NULL)
}
\arguments{
\item{data_dict_list}{A list of data frame(s) representing metadata to be
transformed.}

\item{name_group}{A character string of one column in the dataset that can be
taken as a grouping column.}
}
\value{
A list of data frame(s) identifying a workable data dictionary structure.
}
\description{
Binds a list of data dictionaries into one data dictionary.
This is a wrapper function analogous to \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}}.
}
\details{
A data dictionary contains the list of variables in a dataset and metadata
about the variables and can be associated with a dataset. A data dictionary
object is a list of data frame(s) named 'Variables' (required) and
'Categories' (if any). To be usable in any function, the data frame
'Variables' must contain at least the \code{name} column, with all unique and
non-missing entries, and the data frame 'Categories' must contain at least
the \code{variable} and \code{name} columns, with unique combination of
\code{variable} and \code{name}.
}
\seealso{
\code{\link[dplyr:bind_rows]{dplyr::bind_rows()}}
}
