% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{data_dict_match_dataset}
\alias{data_dict_match_dataset}
\title{Inner join between any data and data dict variables}
\usage{
data_dict_match_dataset(
  data,
  data_dict,
  data_dict_apply = FALSE,
  output = c("data", "data_dict")
)
}
\arguments{
\item{data}{A tibble identifying the input data observations.}

\item{data_dict}{A list of tibble(s) representing meta data.}

\item{data_dict_apply}{whether to apply the data dictionary to its dataset.
The resulting tibble will have for each column its associated meta data as
attributes. The factors will be preserved. FALSE by default.}

\item{output}{A vector of character string which indicates if the function
returns a dataset ('dataset'), data dictionary ('data_dict') of both.
Default is c('dataset','data_dict').}
}
\value{
Either a tibble, identifying the dataset, or a list of tibble(s)
identifying a data dictionary. Returns both in a list by defaut.
}
\description{
Proceeds an inner jointure between any dataset and its assocated
data dictionary.
It returns the user either the sub-dataset or the sub-data dictionary, or
both in a list.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A dataset must be a data frame or data frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data
dictionary can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research stardards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
