% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{data_dict_pivot_longer}
\alias{data_dict_pivot_longer}
\title{Turn column(s) in a data dictionary component into their longer equivalent}
\usage{
data_dict_pivot_longer(data_dict, taxonomy)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be
transformed.}

\item{taxonomy}{A data frame or data frame extension (e.g. a tibble),
identifying the scheme used for variables classification as a tibble.}
}
\value{
A list of tibble(s) identifying a data dictionary.
}
\description{
Transforms column(s) in a data dictionary component into their longer
equivalent. If any Opal taxonomy is provided, the correspondent columns in
the data dictionary will be converted to a format that contains fewer columns
than an Opal format to work with. This operation is equivalent to perform a
\code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}} to these columns following the taxonomy structure
provided. The data dictionaries data dictionary provided must respect certain
constrains such as name uniqueness across the variables names.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A taxonomy must be a data frame or data frame extension (e.g. a tibble).
The taxonomy must be compatible with (and generally extracted from) an
Opal environment, and must contain at least 'taxonomy', 'vocabulary' and
'terms' to work with some specific functions. In addition, the taxonomy
may follow Maelstrom research standards, and its content can be evaluated
accordingly, such as naming convention restriction, tagging elements,
or scales, which are specific to Maelstrom Research. In this particular
case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale',
'vocabulary_scale' and 'term_scale' to work with some specific functions.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}, \code{\link[=as_data_dict]{as_data_dict()}}
}
