% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data_visualize.R
\name{dataset_visualize}
\alias{dataset_visualize}
\title{Generate a web-based bookdown visual report of a dataset}
\usage{
dataset_visualize(
  dataset = tibble(id = as.character()),
  bookdown_path,
  data_dict = data_dict_extract(dataset),
  group_by = NULL,
  taxonomy = NULL,
  valueType_guess = FALSE,
  dataset_name = .dataset_name,
  dataset_summary = .summary_var,
  .summary_var = NULL,
  .dataset_name = NULL
)
}
\arguments{
\item{dataset}{A data frame identifying the input dataset observations
associated to its data dictionary.}

\item{bookdown_path}{A character string identifying the folder path where
the bookdown report files will be saved.}

\item{data_dict}{A list of data frame(s) representing meta data of an
associated dataset. Automatically generated if not provided.}

\item{group_by}{A character string identifying the column in each
dataset to use as a grouping variable. Visual elements will be grouped by
this column.}

\item{taxonomy}{An optional data frame identifying a variable
classification schema.}

\item{valueType_guess}{Whether the output should be generated based on more
precise valueType inferred from the data. FALSE by default
(will use the valueType declared).}

\item{dataset_name}{A character string specifying the name of the dataset
(used for internal processes and programming).}

\item{dataset_summary}{A list which identifies an existing
summary produced by \code{\link[=dataset_summarize]{dataset_summarize()}} of the dataset.
Using this parameter can save time in generating the visual report.}

\item{.summary_var}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{.dataset_name}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\value{
A bookdown folder containing files in the specified output folder. To
open the file in browser, open 'docs/index.html'. Or use
\code{\link[=bookdown_open]{bookdown_open()}}
}
\description{
Generates a visual report for a dataset in an HTML bookdown document. The
report provides figures and descriptive statistics for each variable to
facilitate the assessment of input data. Statistics and figures are
generated according to variable data type. The report can be used to help
assess data structure, coherence across elements, and taxonomy or
data dictionary formats. The summaries and figures provide additional
information about variable distributions and descriptive statistics.
The charts and tables are produced based on their data type. The variable
can be grouped using \code{group_by} parameter, which is a (categorical) column
in the dataset. The user may need to use \code{\link[=as.factor]{as.factor()}} in this context. To
fasten the process (and allow recycling object in a workflow) the user can
feed the function with a \code{dataset_summary}, which is the output of the function
\code{\link[=dataset_summarize]{dataset_summarize()}} of the column(s) \code{col} and  \code{group_by}. The summary
must have the same parameters to operate.
}
\details{
A dataset is a data table containing variables. A dataset object is a
data frame and can be associated with a data dictionary. If no
data dictionary is provided with a dataset, a minimum workable
data dictionary will be generated as needed within relevant functions. An
identifier variable(s) for indexing can be specified by the user.
The id values must be non-missing and will be used in functions that
require it. If no identifier variable is specified, indexing is
handled automatically by the function.

A data dictionary contains the list of variables in a dataset and metadata
about the variables and can be associated with a dataset. A data dictionary
object is a list of data frame(s) named 'Variables' (required) and
'Categories' (if any). To be usable in any function, the data frame
'Variables' must contain at least the \code{name} column, with all unique and
non-missing entries, and the data frame 'Categories' must contain at least
the \code{variable} and \code{name} columns, with unique combination of
\code{variable} and \code{name}.

The valueType is a declared property of a variable that is required in
certain functions to determine handling of the variables. Specifically,
valueType refers to the
\href{https://opaldoc.obiba.org/en/variables-data.html#value-types}{OBiBa data type of a variable}.
The valueType is specified in a data dictionary in a column 'valueType' and
can be associated with variables as attributes. Acceptable valueTypes
include 'text', 'integer', 'decimal', 'boolean', datetime', 'date'. The full
list of OBiBa valueType possibilities and their correspondence with R data
types are available using \link{valueType_list}.

A taxonomy is a classification schema that can be defined for variable
attributes. A taxonomy is usually extracted from an
\href{https://www.obiba.org/pages/products/opal/}{Opal environment}, and a
taxonomy object is a data frame that must contain at least the columns
\code{taxonomy}, \code{vocabulary}, and \code{terms}. Additional details about Opal
taxonomies are
\href{https://opaldoc.obiba.org/en/latest/web-user-guide/administration/taxonomies.html}{available online}.
}
\examples{
{

# You can use our demonstration files to run examples

library(fs)

dataset <- madshapR_DEMO$dataset_TOKYO['height']
dataset_summary <- madshapR_DEMO$`dataset_summary`

if(dir_exists(tempdir())) dir_delete(tempdir())
bookdown_path <- tempdir()

dataset_visualize(
  dataset,
  dataset_summary = dataset_summary,
  bookdown_path = bookdown_path)
  
# To open the file in browser, open 'bookdown_path/docs/index.html'. 
# Or use bookdown_open(bookdown_path) function.

}

}
\seealso{
\code{\link[=bookdown_open]{bookdown_open()}}
}
