% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataset_functions.R
\name{dataset_zap_data_dict}
\alias{dataset_zap_data_dict}
\title{Remove value labels of a data frame, leaving its unlabelled columns.}
\usage{
dataset_zap_data_dict(dataset)
}
\arguments{
\item{dataset}{A tibble identifying the input data observations associated to
its data dictionary.}
}
\value{
A tibble identifying a dataset.
}
\description{
Removes any attributes attached to a tibble. Any value in columns will be preserved.
}
\details{
A dataset must be a data frame or data frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data
dictionary can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research stardards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[haven:zap_labels]{haven::zap_labels()}}.
}
