% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{DEMO_files}
\alias{DEMO_files}
\title{Built-in material allowing the user to test the package with demo data}
\format{
\subsection{\code{list}}{

A list with 20 elements used for testing the package
(data frames and lists):

\describe{
\item{dd_MELBOURNE_1_format_maelstrom}{Data dictionary (1) of Melbourne
dataset}
\item{dd_MELBOURNE_2_format_maelstrom}{Data dictionary (2) of Melbourne
dataset}
\item{dd_PARIS_format_flatten - ERROR}{Data dictionary of Paris dataset
containing errors for testing purpose}
\item{dd_PARIS_format_flatten}{Data dictionary of Paris in preprocessed
format}
\item{dd_PARIS_format_maelstrom}{Data dictionary of Paris dataset}
\item{dd_TOKYO_format_maelstrom}{Data dictionary of Tokyo dataset}
\item{dd_TOKYO_format_maelstrom_tagged - ERROR WITH DATA}{Tagged
data dictionary of Tokyo dataset containing errors for testing purpose}
\item{dd_TOKYO_format_maelstrom_tagged - ERROR}{Tagged data dictionary of
Tokyo dataset containing errors for testing purpose}
\item{dd_TOKYO_format_maelstrom_tagged}{Tagged data dictionary of Tokyo
dataset}
\item{dd_TOKYO_format_opal_tagged - ERROR WITH TAXO}{Tagged
data dictionary of Tokyo dataset containing errors for testing purpose}
\item{dd_TOKYO_format_opal_tagged}{Tagged data dictionary of Tokyo dataset
opal format}
\item{dataset_MELBOURNE_1}{Dataset of Melbourne (1)}
\item{dataset_MELBOURNE_2}{Dataset of Melbourne (2)}
\item{dataset_PARIS}{Dataset of Paris}
\item{dataset_TOKYO - ERROR WITH DATA}{Dataset of Tokyo containing errors
for testing purpose}
\item{dataset_TOKYO}{Dataset of Tokyo}
\item{taxonomy_PARIS}{Taxonomy specific to Paris dataset}
\item{taxonomy_opal}{Opal Taxonomy}
\item{taxonomy_opal_mlstr}{Maelstrom Taxonomy}
\item{summary_var}{Variables summary for testing purpose}
...
}
}
}
\usage{
DEMO_files
}
\description{
Built-in tibbles and lists allowing the user to test the package with demo
material.
}
\examples{
{

 print(DEMO_files$dataset_TOKYO)

}
}
\keyword{datasets}
