% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataset_functions.R
\name{dossier_create}
\alias{dossier_create}
\title{Create a dossier object from a list of dataset(s)}
\usage{
dossier_create(dataset_list, data_dict_apply = FALSE)
}
\arguments{
\item{dataset_list}{A list of data frame(s), identifying the input data
observations.}

\item{data_dict_apply}{whether to apply the data dictionary to its dataset.
The resulting data frame will have for each column its associated meta data
as attributes. FALSE by default.}
}
\value{
A list of data frame(s), each of them identifying datasets in a dossier.
}
\description{
Assembles a dossier object from the listed datasets. A dossier is a list
containing at least one valid dataset and is the input used by key functions
of the package.
}
\details{
A dataset is a data table containing variables. A dataset object is a
data frame and can be associated with a data dictionary. If no
data dictionary is provided with a dataset, a minimum workable
data dictionary will be generated as needed within relevant functions. An
identifier variable(s) for indexing can be specified by the user.
The id values must be non-missing and will be used in functions that
require it. If no identifier variable is specified, indexing is
handled automatically by the function.
}
\examples{
{

# use madshapR_DEMO provided by the package

###### Example 1: datasets can be gathered into a dossier which is a list.
dossier <- dossier_create(
 dataset_list = list(
   dataset_MELBOURNE = madshapR_DEMO$dataset_MELBOURNE,
   dataset_PARIS = madshapR_DEMO$dataset_PARIS ))
    
###### Example 2: any data frame can be gathered into a dossier
dossier_create(list(iris, mtcars))
   
}

}
