% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data_visualize.R
\name{identify_visual_type}
\alias{identify_visual_type}
\title{Identify visual type of a variable based on valueType}
\usage{
identify_visual_type(dataset, data_dict)
}
\arguments{
\item{dataset}{A tibble identifying the input data observations associated to
its data dictionary.}

\item{data_dict}{A list of tibble(s) representing meta data of an
associated dataset. Automatically generated if not provided.}
}
\value{
A list of two tibbles which makes up the data dictionary in
Maelstrom Research format where a column 'viz_type' has been added to the
data dictionary provided as an input.
}
\description{
This helper function analyses the content of a dataset and its data dictionary
to extract the type of visualization to generate in a report. This function can
be used to manually personalize the report parameters.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A dataset must be a data frame or data frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data
dictionary can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research stardards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
\dontrun{
# Example 1: viz type of iris dataset
library(tidyverse)
identify_visual_type(dataset = iris) \%>\% .$Variables \%>\% select(name,viz_type)
}

}
