% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{is_mlstr_data_dict}
\alias{is_mlstr_data_dict}
\title{Evaluate if any object is a Maelstrom formatted data dictionary or not}
\usage{
is_mlstr_data_dict(object)
}
\arguments{
\item{object}{A potential Maelstrom formatted data dictionary to be
evaluated.}
}
\value{
A logical.
}
\description{
Confirms whether the input object is a valid data dictionary compliant with
Maelstrom Research format. This function mainly helps validate input within
other functions of the package but could be used to check if a taxonomy is
valid.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
\dontrun{
# example
}

}
\seealso{
For a better assessment, please use \code{\link[=data_dict_evaluate]{data_dict_evaluate()}}.
}
