% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataset_functions.R
\name{is_study}
\alias{is_study}
\title{Evaluate if any object is a study or not}
\usage{
is_study(object)
}
\arguments{
\item{object}{A potential study to be evaluated.}
}
\value{
A logical.
}
\description{
Confirms whether the input object is a valid study.
This function mainly helps validate input within other functions of the
package but could be used to check if a study is valid.
}
\details{
A study must be a named list containing at least one data frame or
data frame extension (e.g. a tibble), each of them being datasets.
The name of each tibble will be use as the reference name of the dataset.
}
\examples{
\dontrun{
# example
}

}
