% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-catalogue_functions.R
\name{is_taxonomy}
\alias{is_taxonomy}
\title{Evaluate if any object is a taxonomy scheme or not}
\usage{
is_taxonomy(object)
}
\arguments{
\item{object}{A potential taxonomy to be evaluated.}
}
\value{
A logical.
}
\description{
Confirms whether the input object is a valid taxonomy.
This function mainly helps validate input within other functions of the
package but could be used to check if a taxonomy is valid.
}
\details{
A taxonomy must be a data frame or data frame extension (e.g. a tibble).
The taxonomy must be compatible with (and generally extracted from) an
Opal environment, and must contain at least 'taxonomy', 'vocabulary' and
'terms' to work with some specific functions. In addition, the taxonomy
may follow Maelstrom research standards, and its content can be evaluated
accordingly, such as naming convention restriction, tagging elements,
or scales, which are specific to Maelstrom Research. In this particular
case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale',
'vocabulary_scale' and 'term_scale' to work with some specific functions.
}
\examples{
\dontrun{
# example
}

}
