% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-catalogue_functions.R
\name{is_valueType}
\alias{is_valueType}
\title{Evaluate if any object is a valid valueType name or not}
\usage{
is_valueType(object)
}
\arguments{
\item{object}{A potential valueType name to be evaluated.}
}
\value{
A logical.
}
\description{
Uses a data dictionary in the Maelstrom Research formats (with 'Variables'
and 'Categories' in separate tibbles and standard columns in each) to apply
their valueType to a dataset in tibble format. If no data dictionary is
provided, the function will automatically evaluate the most restrictive
valueType for each variable in the dataset and apply it.
}
\details{
valueType is one of the property of a variable entity. It refers to
the (Obiba internal) type of any variable. The valueType can be 'text',
'integer', 'decimal', 'boolean', 'locale', 'datetime', 'date', 'binary',
'point', 'linestring', 'polygon'.
The valueType list is available using \link{valueType_list}, and their
corresponding with typeof which is the (R internal) type of any object.
}
\examples{
\dontrun{
# example
}

}
\seealso{
\href{https://opaldoc.obiba.org/en/dev/magma-user-guide/value/type.html}{Opal documentation}
}
