% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madshapR-package.R
\docType{package}
\name{madshapR-package}
\alias{madshapR}
\alias{madshapR-package}
\title{madshapR: Functions to Support Technical Processes and Data Processing with Maelstrom Research tools}
\description{
Functions to support data cleaning, evaluation, and description, developed for integration with Maelstrom Research software tools. 'madshapR' provides functions primarily to evaluate and manipulate datasets and data dictionaries in preparation for data harmonization with the package 'Rmonize' and to facilitate integration and transfer between RStudio servers and secure Opal environments. 'mapshapR' functions can be used independently but are optimized in conjunction with ‘Rmonize’ functions for streamlined and coherent harmonization processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maelstrom-research/madshapR}
  \item Report bugs at \url{https://github.com/maelstrom-research/madshapR/issues}
}

}
\author{
\strong{Maintainer}: Guillaume Fabre \email{guijoseph.fabre@gmail.com} (\href{https://orcid.org/0000-0002-0124-9970}{ORCID})

Other contributors:
\itemize{
  \item Maelstrom-research group \email{info@maelstrom-research.org} [funder, copyright holder]
}

}
\keyword{internal}
