% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{madshapR_DEMO}
\alias{madshapR_DEMO}
\title{Built-in material allowing the user to test the package with demo data}
\format{
\subsection{\code{list}}{

A list with 20 elements used for testing the package
(data frames and lists):

\describe{
\item{data_dict_MELBOURNE}{Data dictionary of Melbourne
dataset}
\item{data_dict_PARIS}{Data dictionary of Paris dataset}
\item{data_dict_PARIS - flatten}{Data dictionary of Paris in preprocessed
format}
\item{data_dict_TOKYO}{Data dictionary of Tokyo dataset}
\item{data_dict_TOKYO - with errors}{Tagged data dictionary of
Tokyo dataset containing errors for testing purpose}
\item{data_dictt_TOKYO - with errors in data}{data dictionary of Tokyo dataset
containing errors with for testing purpose}
\item{dataset_MELBOURNE}{Dataset of Melbourne}
\item{dataset_PARIS}{Dataset of Paris}
\item{dataset_TOKYO}{Dataset of Tokyo}
\item{data_dict_TOKYO - with errors in data}{Dataset of Tokyo containing errors
for testing purpose}
...
}
}
}
\usage{
madshapR_DEMO
}
\description{
Built-in tibbles and lists allowing the user to test the package with demo
material.
}
\examples{
{

 print(madshapR_DEMO$dataset_TOKYO)

}
}
\keyword{datasets}
