% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{madshapR_examples}
\alias{madshapR_examples}
\title{Built-in material allowing the user to test the package with example data}
\format{
\subsection{\code{list}}{

A list with 9 elements (data frames and lists) providing example objects
for testing the package:

\describe{
\item{dataset_example}{Dataset for example dataset}
\item{data_dict_example - as_data_dict}{Data dictionary for example
dataset where the structure is a data dictionary}
\item{data_dict_example - as_data_dict_mlstr}{Data dictionary for
example dataset where the structure is a data dictionary compliant with
Maelstrom}
\item{dataset_example - errors with data}{Dataset of example
with errors with example data dictionary}
\item{data_dict_example - errors with data}{Data Dictionary for example
dataset with errors with example dataset}
\item{data_dict_example - errors}{Data dictionary for example dataset
with errors}
\item{data_dict_example - collapsed}{Data dictionary for example
with collapsed categories}
\item{taxonomy_example}{Taxonomy for example dataset}
\item{summary - dataset_example}{Dataset example summary}
...
}
}
}
\usage{
madshapR_examples
}
\description{
Example datasets and data dictionaries, and taxonomy, to provide illustrative
examples of objects used by madshapR.
}
\examples{
{
 
library(dplyr)
 
head(madshapR_examples$`dataset_example`)
glimpse(madshapR_examples$`data_dictionary_example`)

}
}
\keyword{datasets}
