% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{madshapR_help}
\alias{madshapR_help}
\title{Call the help center for full documentation}
\usage{
madshapR_help()
}
\description{
This feature is a direct call of the documentation in the repository hosting
the package. The user accesses the description of the latest version of the
package, the vignettes, and the list of functions.
}
\examples{
{

library(utils)
madshapR_help()

}

}
