% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\name{madshapR_website}
\alias{madshapR_website}
\title{Call to online documentation}
\usage{
madshapR_website()
}
\value{
Nothing to be returned. The function opens a web page.
}
\description{
Direct call to the online documentation for the package, which includes a
description of the latest version of the package, vignettes, user guides,
and a reference list of functions and help pages.
}
\examples{
{

madshapR_website()

}

}
