% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-opal_functions.R
\name{opal_mlstr_taxonomy_get}
\alias{opal_mlstr_taxonomy_get}
\title{Get the taxonomy as used by Maelstrom Research}
\usage{
opal_mlstr_taxonomy_get(opal = NULL)
}
\arguments{
\item{opal}{Opal login attributes.}
}
\value{
A tibble identifying a taxonomy (generally generated from Opal taxonomy.
}
\description{
Downloads all taxonomies from an Opal server in a tibble format. This
taxonomy is specifically the Maelstrom taxonomy used for any purpose of data
transformation, cleaning or assessment, following Maelstrom standards.
}
\details{
The user must be allowed to interact with their Opal. The errors
provided may be associated to the handler or the read/write
permissions to Opal. The user must have adequate credentials
to interact with their Opal environment.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
Please see \href{https://opaldoc.obiba.org/en/dev/}{Opal documentation}for further
information.
}
