% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-opal_functions.R
\name{opal_project_create}
\alias{opal_project_create}
\title{Create a project in an Opal environment}
\usage{
opal_project_create(opal, project, tag = NULL, ...)
}
\arguments{
\item{opal}{Opal login attributes.}

\item{project}{A character string to name the project in Opal.}

\item{tag}{A character string to provide a tag for the Opal project.}

\item{...}{Additional parameters.}
}
\value{
A project in an Opal environment. If the project already exists, it
will remain as it is, and no new project is created.
The user must have adequate credentials to interact with their Opal
environment.
}
\description{
Creates an empty project in Opal environment. It is a wrapper
of \code{\link[opalr:opal.project_create]{opalr::opal.project_create()}}.
}
\details{
The user must be allowed to interact with their Opal. The errors
provided may be associated to the handler or the read/write
permissions to Opal. The user must have adequate credentials
to interact with their Opal environment.
}
\examples{
{

library(opalr)
opal <- 
  opal.login('administrator','password', url ='https://opal-demo.maelstrom-research.org')
  
tempdir <- basename(tempdir())
try(opal_project_create(opal, tempdir))

}

}
\seealso{
Please see \href{https://opaldoc.obiba.org/en/dev/}{Opal documentation}for further
information.
}
