% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-opal_functions.R
\name{opal_taxonomy_get}
\alias{opal_taxonomy_get}
\title{Get the taxonomy as used by Opal}
\usage{
opal_taxonomy_get(opal)
}
\arguments{
\item{opal}{Opal login attributes.}
}
\value{
A tibble identifying a taxonomy (generally generated from Opal taxonomy).
}
\description{
Downloads all taxonomies from an Opal server in the format used by Opal.
This taxonomy is used for any purpose of data transformation, cleaning
or assessement.
}
\details{
The user must be allowed to interact with their Opal. The errors
provided may be associated to the handler or the read/write
permissions to Opal. The user must have adequate credentials
to interact with their Opal environment.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
Please see \href{https://opaldoc.obiba.org/en/dev/}{Opal documentation}for further
information.
}
