% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-import-from-fabR.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{bookdown_template}
\alias{bookdown_render}
\alias{bookdown_open}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fabR}{\code{\link[fabR]{bookdown_open}}, \code{\link[fabR]{bookdown_render}}, \code{\link[fabR]{bookdown_template}}}
}}

