% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataset_functions.R
\name{study_create}
\alias{study_create}
\title{Create a study as a list of datasets}
\usage{
study_create(dataset_list, data_dict_apply = FALSE)
}
\arguments{
\item{dataset_list}{A list of tibble(s), identifying the input data observations.}

\item{data_dict_apply}{whether to apply the data dictionary to its dataset.
The resulting tibble will have for each column its associated meta data as
attributes. The factors will be preserved. FALSE by default.}
}
\value{
A list of tibble(s), each of them identifying datasets in a study.
}
\description{
Assembles a study from datasets. A study is a list containing at least one
dataset, that can be used directly in keys functions of the package.
}
\details{
A dataset must be a data frame or data frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data
dictionary can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research stardards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
