% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-data_summarise.R
\name{summary_variables_numeric}
\alias{summary_variables_numeric}
\title{Provide descriptive statistics for variables of type 'numeric' in a dataset}
\usage{
summary_variables_numeric(
  dataset = NULL,
  data_dict = NULL,
  dataset_preprocess = .dataset_preprocess,
  .dataset_preprocess = NULL
)
}
\arguments{
\item{dataset}{A data frame identifying the input dataset observations
associated to its data dictionary.}

\item{data_dict}{A list of data frame(s) representing meta data of an
associated dataset. Automatically generated if not provided.}

\item{dataset_preprocess}{A data frame which provides summary of the
variables (for internal processes and programming).}

\item{.dataset_preprocess}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\value{
A data frame providing statistical description of 'numerical' variables
present in a dataset.
}
\description{
Summarizes (in a data frame) the columns of type 'numeric' in a dataset and
its data dictionary (if any). The summary provides information about quality,
type, composition, and descriptive statistics of variables. Statistics are
generated by valueType.
}
\details{
A data dictionary contains the list of variables in a dataset and metadata
about the variables and can be associated with a dataset. A data dictionary
object is a list of data frame(s) named 'Variables' (required) and
'Categories' (if any). To be usable in any function, the data frame
'Variables' must contain at least the \code{name} column, with all unique and
non-missing entries, and the data frame 'Categories' must contain at least
the \code{variable} and \code{name} columns, with unique combination of
\code{variable} and \code{name}.

A dataset is a data table containing variables. A dataset object is a
data frame and can be associated with a data dictionary. If no
data dictionary is provided with a dataset, a minimum workable
data dictionary will be generated as needed within relevant functions. An
identifier variable(s) for indexing can be specified by the user.
The id values must be non-missing and will be used in functions that
require it. If no identifier variable is specified, indexing is
handled automatically by the function.
}
\examples{
{

###### Example : any data frame can be a dataset by definition.
dataset_preprocess <- dataset_preprocess(iris)
summary_variables_numeric(dataset_preprocess = dataset_preprocess)

}

}
