% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-data_summarise.R
\name{summary_variables_text}
\alias{summary_variables_text}
\title{Provide statistical description of 'text' variables of a dataset}
\usage{
summary_variables_text(dataset = NULL, data_dict = NULL, .resume_var = NULL)
}
\arguments{
\item{dataset}{A tibble identifying the input dataset observations associated to
its data dictionary.}

\item{data_dict}{A list of tibble(s) representing meta data of an
associated dataset. Automatically generated if not provided.}

\item{.resume_var}{A tibble which provides summary of the variables (for
development purpose only)}
}
\value{
A tibble providing statistical description of 'text' variables present
in a dataset.
}
\description{
Generates a tibble that summarises all 'text' columns
in a dataset with (if any) its data dictionary. The data dictionary (if
present separates observations between open values, missing values,
categorical values, and categorical missing values (which corresponds to the
'missing' column in the Categories' sheet). Statistics are generated
according to their valueType which can be 'text', 'datetime', linestring',
'point', 'locale', or 'polygon'. This summary can be used to assist the user
in the assessment of the dataset composition, with observation distribution
and descriptive statistics.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data-frame or data-frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A dataset must be a data-frame or data-frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data dictionary
can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research standards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
{

# use DEMO_files provided by the package
library(dplyr)

###### Example 1: Combine functions and summarise datasets.
data_dict <- as_data_dict_mlstr(DEMO_files$dd_TOKYO_format_maelstrom_tagged)
dataset <-
  DEMO_files$dataset_TOKYO \%>\%
  valueType_adjust(from = data_dict) \%>\%
  data_dict_apply(data_dict) \%>\%
  select(where(is.character))
  
.resume_var <- resume_variables(dataset)
summary_variables_text(.resume_var = .resume_var)
  
###### Example 2: any data-frame (or tibble) can be a dataset by definition.
.resume_var <- 
  resume_variables(
    mutate(iris['Species'],Species = as.character(Species)))
summary_variables_text(.resume_var = .resume_var)

}

}
