% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-opal_functions.R
\name{taxonomy_opal_mlstr_get}
\alias{taxonomy_opal_mlstr_get}
\title{Get the taxonomy as used by Maelstrom Research}
\usage{
taxonomy_opal_mlstr_get(opal = NULL)
}
\arguments{
\item{opal}{Opal login attributes.}
}
\value{
A tibble identifying a taxonomy (generally generated from Opal taxonomy.
}
\description{
Downloads all taxonomies from an Opal server in a tibble format. This
taxonomy is specifically the Maelstrom taxonomy used for any purpose of dataset
transformation, cleaning or assessment, following Maelstrom standards.
}
\details{
The user must be allowed to interact with their Opal. The errors
provided may be associated to the handler or the read/write
permissions to Opal. The user must have adequate credentials
to interact with their Opal environment.
}
\examples{
{

library(opalr)
opal <- 
  opal.login('administrator','password', 
    url ='https://opal-demo.maelstrom-research.org/ui/index.html')
  
try(taxonomy_opal_mlstr_get(opal))
  
}

}
\seealso{
Please see \href{https://opaldoc.obiba.org/en/dev/}{Opal documentation}for further
information.
}
