% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-catalogue_functions.R
\name{valueType_of}
\alias{valueType_of}
\title{Give the valueType of any object}
\usage{
valueType_of(x)
}
\arguments{
\item{x}{R object. Can be a vector.}
}
\value{
A character string which is the valueType of the given object.
}
\description{
Determines the valueType based on \code{\link[base:typeof]{base::typeof()}} and \code{\link[base:class]{base::class()}} of
an object.
The possible values returned are 'date', 'boolean', 'integer',
'decimal', and 'text'.
}
\details{
valueType is one of the property of a variable entity. It refers to
the (Obiba internal) type of any variable. The valueType can be 'text',
'integer', 'decimal', 'boolean', 'locale', 'datetime', 'date', 'binary',
'point', 'linestring', 'polygon'.
The valueType list is available using \link{valueType_list}, and their
corresponding with typeof which is the (R internal) type of any object.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[base:typeof]{base::typeof()}}, \code{\link[base:class]{base::class()}}
\link{valueType_list} for insights about possible valueType and
translation into type and class in R.
\href{https://opaldoc.obiba.org/en/dev/magma-user-guide/value/type.html}{Opal documentation}
}
