% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data_visualize.R
\name{variable_visualize}
\alias{variable_visualize}
\title{Generate a combined plot of a variable in a dataset}
\usage{
variable_visualize(
  dataset = tibble(id = as.character()),
  col,
  data_dict = NULL,
  group_by = NULL,
  .summary_var = NULL
)
}
\arguments{
\item{dataset}{A tibble identifying the input dataset observations associated to
its data dictionary.}

\item{col}{A character string specifying the name of the column.}

\item{data_dict}{A list of tibble(s) representing meta data of an
associated dataset. Automatically generated if not provided.}

\item{group_by}{A character string of one column in the dataset that can be
taken as a grouping column. The visual element will be grouped and displayed
by this column.}

\item{.summary_var}{A summary list which provides summary of the variables (
If provided, the function will not do this action internally, enhancing the
speed of completion)}
}
\value{
xxx xxx xxx
}
\description{
xxx xxx xxx
}
\details{
A data dictionary-like structure must be a list of at least one or two
data-frame or data-frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A dataset must be a data-frame or data-frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data dictionary
can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research standards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
{

variable_visualize(dataset = iris, col = Petal.Length, group_by = Species)

}

}
\seealso{
\code{\link[=open_visual_report]{open_visual_report()}}
}
