% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{as_data_dict}
\alias{as_data_dict}
\title{Validate and coerce any object as data dictionary}
\usage{
as_data_dict(object)
}
\arguments{
\item{object}{A potential valid data dictionary to be coerced.}
}
\value{
A list of tibble(s) identifying a data dictionary.
}
\description{
Confirms that the input object is a valid data dictionary, and return it as a
data dictionary with the appropriate mlstr_class attribute. This function
mainly helps validate input within other functions of the package but could
be used to check if a data dictionary is valid.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
{

# use DEMO_files provided by the package

data_dict <- DEMO_files$dd_PARIS_format_maelstrom
as_data_dict(data_dict)

}

}
\seealso{
For a better assessment, please use \code{\link[=data_dict_evaluate]{data_dict_evaluate()}}.
}
