% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataset_functions.R
\name{as_dataset}
\alias{as_dataset}
\title{Validate and coerce any object as dataset}
\usage{
as_dataset(object, col_id = NULL)
}
\arguments{
\item{object}{A potential dataset to be coerced.}

\item{col_id}{A character string specifying the name(s) of the column(s)
which refer to key identifier of the dataset. The column(s) must be named
of indicated by position.}
}
\value{
A tibble identifying a dataset.
}
\description{
Confirms that the input object is a valid dataset, and return it as a dataset
with the appropriate mlstr_class attribute. This function mainly helps
validate input within other functions of the package but could be used to
check if a dataset is valid.
}
\details{
A dataset must be a data frame or data frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data
dictionary can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research standards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
{

# use DEMO_files provided by the package

###### Example 1: a dataset can have an id column(s) which is specified as
# an attribute. 
dataset <- DEMO_files$dataset_MELBOURNE_1
as_dataset(dataset, col_id = "id")

###### Example 2: any data-frame (or tibble) can be a dataset by definition.
as_dataset(iris)
}

}
