% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataset_functions.R
\name{as_study}
\alias{as_study}
\title{Validate and coerce any object as study}
\usage{
as_study(object)
}
\arguments{
\item{object}{A potential study to be coerced.}
}
\value{
A list of tibble(s), each of them identifying datasets in a study.
}
\description{
Confirms that the input object is a valid study, and return it as a study with
the appropriate mlstr_class attribute. This function mainly helps validate input
within other functions of the package but could be used to check if a study is
valid.
}
\details{
A study must be a named list containing at least one data frame or
data frame extension (e.g. a tibble), each of them being datasets.
The name of each tibble will be use as the reference name of the dataset.
}
\examples{
\dontrun{
# example
}

}
\seealso{
For a better assessment, please use \code{\link[=dataset_evaluate]{dataset_evaluate()}}.
}
