% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-catalogue_functions.R
\name{as_taxonomy}
\alias{as_taxonomy}
\title{Validate and coerce any object as taxonomy}
\usage{
as_taxonomy(object)
}
\arguments{
\item{object}{A potential taxonomy to be coerced.}
}
\value{
A tibble identifying a taxonomy (generally generated from Opal taxonomy).
}
\description{
Confirms that the input object is a valid taxonomy, and return it as a
taxonomy with the appropriate madshapR::class attribute. This function mainly
helps validate input within other functions of the package but could be used
to check if a taxonomy is valid.
}
\details{
A taxonomy must be a data-frame or data-frame extension (e.g. a tibble).
The taxonomy must be compatible with (and generally extracted from) an
Opal environment, and must contain at least 'taxonomy', 'vocabulary' and
'terms' to work with some specific functions. In addition, the taxonomy
may follow Maelstrom research standards, and its content can be evaluated
accordingly, such as naming convention restriction, tagging elements,
or scales, which are specific to Maelstrom Research. In this particular
case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale',
'vocabulary_scale' and 'term_scale' to work with some specific functions.
}
\examples{
{

# use DEMO_files provided by the package

###### Example 1: this function is compatible with taxonomy_opal_get()
library(opalr)
opal <- 
  opal.login('administrator','password',url = 'https://opal-demo.obiba.org/')
opal_taxo <- taxonomy_opal_mlstr_get(opal)
as_taxonomy(opal_taxo)

###### Example 2: you can create your own taxonomy
print(DEMO_files$taxonomy_PARIS)
as_taxonomy(DEMO_files$taxonomy_PARIS)

}

}
\seealso{
\href{https://opaldoc.obiba.org/en/dev/magma-user-guide/value/type.html}{Opal documentation}
\code{\link[=taxonomy_opal_get]{taxonomy_opal_get()}}
}
