% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-catalogue_functions.R
\name{as_valueType}
\alias{as_valueType}
\title{Validate and coerce any object according to a given valueType}
\usage{
as_valueType(x, valueType = "text")
}
\arguments{
\item{x}{R object to be coerced. Can be a vector.}

\item{valueType}{A character string of the valueType used to coerce x.}
}
\value{
The object coerced accordingly to the given valueType.
}
\description{
Uses a data dictionary in the Maelstrom Research formats (with 'Variables'
and 'Categories' in separate tibbles and standard columns in each) to apply
their valueType to a dataset in tibble format. If no data dictionary is
provided, the function will automatically evaluate the most restrictive
valueType for each variable in the dataset and apply it.
}
\details{
valueType is one of the property of a variable entity. It refers to
the (Obiba internal) type of any variable. The valueType can be 'text',
'integer', 'decimal', 'boolean', 'locale', 'datetime', 'date', 'binary',
'point', 'linestring', 'polygon'
The valueType list is available using \link{valueType_list}, and their
corresponding with typeof which is the (R internal) type of any object.
}
\examples{
{

# use DEMO_files provided by the package

dataset <- DEMO_files$dataset_TOKYO
valueType_of(dataset$dob)
valueType_guess(dataset$dob)
as_valueType(dataset$dob,'date') 

# as_valueType is compatible with tidyverse philosophy
library(dplyr)
mtcars \%>\% 
  mutate(BMI = as_valueType(cyl,'integer'))

}

}
\seealso{
\href{https://opaldoc.obiba.org/en/dev/magma-user-guide/value/type.html}{Opal documentation}
}
