% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-unit_checks.R
\name{check_data_dict_missing_categories}
\alias{check_data_dict_missing_categories}
\title{Return categorical values which 'missing' column is not a boolean}
\usage{
check_data_dict_missing_categories(data_dict)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be evaluated.}
}
\value{
A tibble providing categorical values which 'missing' column is
not a boolean.
}
\description{
Generates a tibble that reports any categorical variable where
'missing' column is not a boolean (or compatible value that can be coerced into
a boolean).
This tibble can be used to assist the user in the assessment of the data
structure, fields investigation (mandatory or not), coherence across elements
and taxonomy or standard evaluation.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
