% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-unit_checks.R
\name{check_data_dict_valueType}
\alias{check_data_dict_valueType}
\title{Return evaluation of the valueType declared in a data dictionary}
\usage{
check_data_dict_valueType(data_dict)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be evaluated.}
}
\value{
A tibble providing non-standard valueType declared in a data dictionary.
}
\description{
Generates a tibble that reports any variable whose valueType is
not present in the list of authorized valueType list. Additionally, it
evaluates if the valueType is compatible with categorical values declared.
This tibble can be used to assist the user in the assessment of the dataset
structure, fields investigation (mandatory or not), coherence across elements
and taxonomy or standard evaluation.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data-frame or data-frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
{

# use DEMO_files provided by the package

data_dict <- DEMO_files$`dd_TOKYO_format_maelstrom_tagged - ERROR`
check_data_dict_valueType(data_dict)

}

}
