% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-unit_checks.R
\name{check_data_dict_variables}
\alias{check_data_dict_variables}
\title{Assess a data dictionary for potential issues in variables}
\usage{
check_data_dict_variables(data_dict)
}
\arguments{
\item{data_dict}{A list of data frame(s) representing metadata to be evaluated.}
}
\value{
A data frame providing non unique variables across a data dictionary.
}
\description{
Generates a data frame report of any non-unique variable names in the
'Variables' element.
This report can be used to help assess data structure, presence of fields,
coherence across elements, and taxonomy or data dictionary formats.
}
\details{
A data dictionary contains the list of variables in a dataset and metadata
about the variables and can be associated with a dataset. A data dictionary
object is a list of data frame(s) named 'Variables' (required) and
'Categories' (if any). To be usable in any function, the data frame
'Variables' must contain at least the \code{name} column, with all unique and
non-missing entries, and the data frame 'Categories' must contain at least
the \code{variable} and \code{name} columns, with unique combination of
\code{variable} and \code{name}.
}
