% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-unit_checks.R
\name{check_name_standards}
\alias{check_name_standards}
\title{Return non-standard names across a vector}
\usage{
check_name_standards(var_names)
}
\arguments{
\item{var_names}{A character vector of names.}
}
\value{
A tibble providing non-standard names across a vector.
}
\description{
Generates a tibble that reports any variable whose name is not
standard, as defined by Maelstrom naming standards.
This tibble can be used to assist the user in the assessment of the data
structure, fields investigation (mandatory or not), coherence across elements
and taxonomy or standard evaluation.
}
\details{
The user must provide element which respect a certain structure to work with
the functions of the package or its environment (Maelstrom and/or
Obiba suite). In addition, any element may follow Maelstrom research
stardards, and its content can be evaluated accordingly, such as naming
convention restriction, columns like 'valueType', , and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
