% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-data_evaluate.R
\name{data_dict_evaluate}
\alias{data_dict_evaluate}
\title{Generate a report as an Excel spreadsheet of a study-specific data dictionary}
\usage{
data_dict_evaluate(data_dict, taxonomy = NULL, as_mlstr_data_dict = TRUE)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be evaluated.}

\item{taxonomy}{A tibble identifying the scheme used for variables
classification as a tibble.}

\item{as_mlstr_data_dict}{Whether the output data dictionary has a simple
data dictionary structure or not (meaning has a Maelstrom data dictionary
structure, compatible with Maelstrom ecosystem such as Opal environment).
FALSE by default.}
}
\value{
A list of tibbles of report for one study-specific data dictionary.
}
\description{
Generates an Excel spreadsheet report for a data dictionary
for each variable to facilitate the assessment of input data.
This report can be used to assist the user in the assessment of the data
structure, fields investigation (mandatory or not), coherence across elements
and taxonomy, or standard evaluation. The summary associated provides data
dictionary composition, and evaluates coherence between elements.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A taxonomy must be a data frame or data frame extension (e.g. a tibble).
The taxonomy must be compatible with (and generally extracted from) an
Opal environment, and must contain at least 'taxonomy', 'vocabulary' and
'terms' to work with some specific functions. In addition, the taxonomy
may follow Maelstrom research stardards, and its content can be evaluated
accordingly, such as naming convention restriction, tagging elements,
or scales, which are specific to Maelstrom Research. In this particular
case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale',
'vocabulary_scale' and 'term_scale' to work with some specific functions.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
dataset <- study_TOKYO \%>\% mutate(dob = fabR::as_any_date(
dob, format = ""mdy""))
data_dict <- dd_TOKYO_format_maelstrom_tagged
study_assessement_report(dataset, data_dict)
}

}
