% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{data_dict_group_by}
\alias{data_dict_group_by}
\title{Group listed data dictionaries using column names}
\usage{
data_dict_group_by(data_dict, col)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be
transformed.}

\item{col}{variable to group by.}
}
\value{
A list of tibble(s) identifying a data dictionary - like structure.
}
\description{
Groups the data dictionary both in the 'Variables' and 'Categories' (assuming
the group is defined in both elements) defined by the query. The user would
group their data dictionary analogously they would do to a tibble with
\code{\link[dplyr:group_by]{dplyr::group_by()}}. \code{\link[=data_dict_ungroup]{data_dict_ungroup()}} reverses the effect of
\link{data_dict_group_by}.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
{

# use DEMO_files provided by the package
# Create a list of data dictionaries where the column 'table' is added to 
# refer to the associated dataset. The object created is not a data 
# data dictionary per say, but can be used as a structure which can be 
# shaped into a data data dictionary.
library(dplyr)

data_dict_list <- 
  DEMO_files[c('dd_MELBOURNE_1_format_maelstrom',
               'dd_MELBOURNE_2_format_maelstrom')] \%>\%
data_dict_list_nest(name_group = 'table')
 
data_dict_group_by(data_dict_list,col = "table")
}

}
\seealso{
\code{\link[dplyr:group_by]{dplyr::group_by()}}, \code{\link[=data_dict_ungroup]{data_dict_ungroup()}}
}
