% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{data_dict_group_split}
\alias{data_dict_group_split}
\title{Split grouped data dictionaries into a named list}
\usage{
data_dict_group_split(data_dict, ...)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be
transformed.}

\item{...}{Column in the data dictionary to split it by. If not provided, the
spliting will be done on the grouping element of a grouped data dictionary.}
}
\value{
A list of tibble(s) identifying a list of data dictionary - like structure.
}
\description{
Divides the data dictionary both in the 'Variables' and 'Categories'
(assuming the group is defined in both elements) into the groups defined by
the query. The list is created analogously the user would do with
\code{\link[dplyr:group_by]{dplyr::group_by()}} and \code{\link[dplyr:group_split]{dplyr::group_split()}}. Each element is named using
the group values. \code{\link[=data_dict_list_nest]{data_dict_list_nest()}} reverses the effect of
\code{\link[=data_dict_group_split]{data_dict_group_split()}}.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[dplyr:group_by]{dplyr::group_by()}}, \code{\link[dplyr:group_split]{dplyr::group_split()}} ,
\code{\link[=data_dict_group_by]{data_dict_group_by()}}, \code{\link[=data_dict_list_nest]{data_dict_list_nest()}}
}
