% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{data_dict_list_nest}
\alias{data_dict_list_nest}
\title{Regroup listed split data dictionaries}
\usage{
data_dict_list_nest(data_dict_list, name_group = NULL)
}
\arguments{
\item{data_dict_list}{A list of tibble(s) representing meta data to be
transformed.}

\item{name_group}{A character string of one column in the dataset that can be
taken as a grouping column.}
}
\value{
A list of tibble(s) identifying a data dictionary - like structure.
}
\description{
Regroups data dictionaries provided in a list, and binds them together.
This function is a wrapper analogous to \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}} from the
tidyverse.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
{

# use DEMO_files provided by the package
# Create a list of data dictionaries where the column 'table' is added to 
# refer to the associated dataset. The object created is not a data 
# data dictionary per say, but can be used as a structure which can be 
# shaped into a data data dictionary.

data_dict_list <- DEMO_files[
    c('dd_MELBOURNE_1_format_maelstrom',
      'dd_MELBOURNE_2_format_maelstrom')]
 
data_dict_list_nest(data_dict_list,name_group = "table")

}

}
\seealso{
\code{\link[dplyr:bind_rows]{dplyr::bind_rows()}}
}
