% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{data_dict_ungroup}
\alias{data_dict_ungroup}
\title{Ungroup data dictionaries}
\usage{
data_dict_ungroup(data_dict)
}
\arguments{
\item{data_dict}{A list of tibble(s) representing meta data to be
transformed.}
}
\value{
A list of tibble(s) identifying a data dictionary - like structure.
}
\description{
Ungroup data dictionaries, both in the 'Variables' and 'Categories' (assuming
the group is defined in both elements). The user would group their
data dictionary analogously they would do to a tibble with
\code{\link[dplyr:group_by]{dplyr::ungroup()}}.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[dplyr:group_by]{dplyr::ungroup()}}
}
