% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataset_functions.R
\name{data_extract}
\alias{data_extract}
\title{Create an empty dataset from a data dictionary}
\usage{
data_extract(data_dict, data_dict_apply = FALSE)
}
\arguments{
\item{data_dict}{A list of data frame(s) representing meta data of an
associated dataset (to be generated).}

\item{data_dict_apply}{Whether to apply the data dictionary to its dataset.
The resulting data frame will have for each column its associated meta data
as attributes. FALSE by default.}
}
\value{
A data frame identifying the dataset created from the variable names list in
'Variables' element of the data dictionary.
}
\description{
Creates an empty dataset using information contained in a data dictionary.
The column names are taken from 'name' in the 'Variables' element of the
data dictionary. If a 'valueType' or alternatively 'typeof' column is
provided, the class of each column is set accordingly (default is text).
}
\details{
A data dictionary contains the list of variables in a dataset and metadata
about the variables and can be associated with a dataset. A data dictionary
object is a list of data frame(s) named 'Variables' (required) and
'Categories' (if any). To be usable in any function, the data frame
'Variables' must contain at least the \code{name} column, with all unique and
non-missing entries, and the data frame 'Categories' must contain at least
the \code{variable} and \code{name} columns, with unique combination of
\code{variable} and \code{name}.
}
\examples{
{

# use madshapR_DEMO provided by the package

data_dict <- madshapR_DEMO$data_dict_MELBOURNE
data_extract(data_dict)

}

}
