% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataset_functions.R
\name{dataset_cat_as_labels}
\alias{dataset_cat_as_labels}
\title{Apply to categorical column(s) their labels declared in the data dictionary}
\usage{
dataset_cat_as_labels(dataset, data_dict = NULL, col_names = names(dataset))
}
\arguments{
\item{dataset}{A tibble identifying the input dataset observations associated to
its data dictionary.}

\item{data_dict}{A list of tibble(s) representing meta data of an
associated dataset (to be generated).}

\item{col_names}{A character string specifying the name(s) of the column(s)
which refer to existing column(s) in the dataset. The column(s) can be named
or indicated by position.}
}
\value{
A tibble identifying a dataset.
}
\description{
Applies to categorical column(s) in a dataset their labels declared in the
categories in the data dictionary.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data-frame or data-frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A dataset must be a data-frame or data-frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data dictionary
can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research standards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
{

# use DEMO_files provided by the package

dataset <- DEMO_files$dataset_TOKYO
data_dict <- as_data_dict_mlstr(DEMO_files$dd_TOKYO_format_maelstrom_tagged)
dataset <- data_dict_apply(dataset,data_dict)
dataset_cat_as_labels(dataset)

}

}
