% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-data_summarise.R
\name{dataset_summarize}
\alias{dataset_summarize}
\title{Generate a report as an Excel spreadsheet of a study-specific dataset}
\usage{
dataset_summarize(
  dataset,
  data_dict = NULL,
  taxonomy = NULL,
  .dataset_name = NULL,
  valueType_guess = FALSE
)
}
\arguments{
\item{dataset}{A tibble identifying the input data observations associated to
its data dictionary.}

\item{data_dict}{A list of tibble(s) representing meta data of an
associated dataset. Automatically generated if not provided.}

\item{taxonomy}{A data frame or data frame extension (e.g. a tibble),
identifying the scheme used for variables classification as a tibble.}

\item{.dataset_name}{A character string specifying the name of the dataset
(internally used in the function \code{datashapR::study_evaluate()}).}

\item{valueType_guess}{Whether the output should include a more accurate
valueType that could be applied to the dataset. FALSE by default.}
}
\value{
A list of tibbles of report for one study-specific data dictionary.
}
\description{
Generates an Excel spreadsheet report for a dataset
for each variable to facilitate the assessment of input data.
This report can be used to assist the user in the assessment of the data
structure, fields investigation (mandatory or not), coherence across elements
and taxonomy, or standard evaluation. The summary associated provides dataset
composition, with observation distribution and descriptive statistics.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A taxonomy must be a data frame or data frame extension (e.g. a tibble).
The taxonomy must be compatible with (and generally extracted from) an
Opal environment, and must contain at least 'taxonomy', 'vocabulary' and
'terms' to work with some specific functions. In addition, the taxonomy
may follow Maelstrom research standards, and its content can be evaluated
accordingly, such as naming convention restriction, tagging elements,
or scales, which are specific to Maelstrom Research. In this particular
case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale',
'vocabulary_scale' and 'term_scale' to work with some specific functions.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
dataset <- study_TOKYO \%>\% mutate(
   dob = fabR::as_any_date(dob, format = ""mdy""))
data_dict <- dd_TOKYO_format_maelstrom_tagged
study_assessment_report(dataset, data_dict)
}

}
\seealso{
\code{\link[=study_evaluate]{study_evaluate()}}
}
