% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data_visualize.R
\name{dataset_visualize}
\alias{dataset_visualize}
\title{Generate a web application (bookdown) report of list of a datasets}
\usage{
dataset_visualize(
  dataset = NULL,
  data_dict = NULL,
  group_by = NULL,
  to,
  out = "ggplot2",
  .keep_files = TRUE
)
}
\arguments{
\item{dataset}{A tibble identifying the input data observations associated to
its data dictionary.}

\item{data_dict}{A list of tibble(s) representing meta data of an
associated dataset. Automatically generated if not provided.}

\item{group_by}{A character string of one column in the dataset that can be
taken as a grouping column. The visual element will be grouped and displayed
by this column.}

\item{to}{A character string identifying the folder path where the bookdown
report will be saved.}

\item{out}{parameter that specifies the graphical outputs expected in the
report. Can be 'ggplot2'.}

\item{.keep_files}{whether to keep the R-markdown files.
FALSE by default.}
}
\value{
A bookdown folder containing files in the specified output folder. To
open the file in browser, open 'index.html'. Or use
\code{\link[=open_visual_report]{open_visual_report()}}
}
\description{
Generates a visual report for a study-specific dataset in an HTML
bookdown document, showing descriptive statistics for each study-specific
variable to facilitate the assessment of input data. Statistics and figures
are generated according to their valueTypes.
This report can be used to assist the user in the assessment of the data
structure, fields investigation (mandatory or not), coherence across elements
and taxonomy or standard evaluation. The summaries and figures associated
provide dataset composition, with observation repartition and descriptive
statistics.
}
\details{
A study must be a named list containing at least one data frame or
data frame extension (e.g. a tibble), each of them being datasets.
The name of each tibble will be use as the reference name of the dataset.
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A dataset must be a data frame or data frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data
dictionary can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research stardards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
\dontrun{
# Create index of files in one folder and read the files
index_DEMO <- file_index_create(folder = ""DEMO"")
file_index_read(index = index_DEMO, file_name = ""study_TOKYO"")
file_index_read(index = index_DEMO, file_name = ""dd_TOKYO"")
# use case 1: report of demo dataset TOKYO
study_visual_report(
dataset = study_TOKYO,
data_dict = dd_TOKYO_format_maelstrom_tagged,
to = ""DEMO/reports/TOKYO"")
# use case 2: report of demo dataset TOKYO, grouped by gndr
study_visual_report(
dataset = study_TOKYO,
data_dict = dd_TOKYO_format_maelstrom_tagged,
to = ""DEMO/reports/TOKYO_gndr"",group_by = ""gndr"",out = ""ggplot2"")
# re-index files to include new files created
index_DEMO <- file_index_create(folder = ""DEMO"")
# read the book down
file_index_read(
 index_DEMO,file_path = ""DEMO/reports/TOKYO_gndr/docs/index.html"")
}

}
\seealso{
\code{\link[=open_visual_report]{open_visual_report()}}
}
