% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataset_functions.R
\name{dossier_create}
\alias{dossier_create}
\title{Create a dossier as a list of datasets}
\usage{
dossier_create(dataset_list, data_dict_apply = FALSE)
}
\arguments{
\item{dataset_list}{A list of tibble(s), identifying the input dataset
observations.}

\item{data_dict_apply}{whether to apply the data dictionary to its dataset.
The resulting tibble will have for each column its associated meta data as
attributes. The factors will be preserved. FALSE by default.}
}
\value{
A list of tibble(s), each of them identifying datasets in a dossier.
}
\description{
Assembles a dossier from datasets. A dossier is a list containing at least one
dataset, that can be used directly in keys functions of the package.
}
\details{
A dataset must be a data-frame or data-frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data dictionary
can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research standards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
{

# use DEMO_files provided by the package

###### Example 1: datasets can be gathered into a dossier which is a list.
dossier <- dossier_create(
 dataset_list = list(
   dataset_MELBOURNE_1 = DEMO_files$dataset_MELBOURNE_1,
   dataset_MELBOURNE_2 = DEMO_files$dataset_MELBOURNE_2))
   
###### Example 2: any data-frame (or tibble) can be gathered into a dossier
dossier_create(list(iris, mtcars))
   
}

}
