% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-data_summarise.R
\name{dossier_summarize}
\alias{dossier_summarize}
\title{Generate a report and summary of a dossier (list of datasets)}
\usage{
dossier_summarize(
  dossier,
  group_by = NULL,
  taxonomy = NULL,
  valueType_guess = FALSE
)
}
\arguments{
\item{dossier}{List of data frame(s), each of them being datasets.}

\item{group_by}{A character string identifying the column in the datasets
to use as a grouping variable. Visual elements will be grouped by this
column.}

\item{taxonomy}{An optional data frame identifying a variable
classification schema.}

\item{valueType_guess}{Whether the output should be generated based on more
precise valueType inferred from the data. FALSE by default
(will use the valueType declared).}
}
\value{
A list of data frames of report for each listed dataset.
}
\description{
Assesses and summarizes the content and structure of a dossier
(list of datasets)  and reports potential issues to facilitate the
assessment of input data. The report can be used to help assess data
structure, presence of fields, coherence across elements, and taxonomy or
data dictionary formats. The summary provides additional information about
variable distributions and descriptive statistics. This report is compatible
with Excel and can be exported as an Excel spreadsheet.
}
\details{
A dossier is a named list containing at least one data frame or more,
each of them being datasets. The name of each tibble will be use as the
reference name of the dataset.

A taxonomy is a classification schema that can be defined for variable
attributes. A taxonomy is usually extracted from an
\href{https://www.obiba.org/pages/products/opal/}{Opal environment}, and a
taxonomy object is a data frame that must contain at least the columns
\code{taxonomy}, \code{vocabulary}, and \code{terms}. Additional details about Opal
taxonomies are
\href{https://opaldoc.obiba.org/en/latest/web-user-guide/administration/taxonomies.html}{available online}.

The valueType is a declared property of a variable that is required in
certain functions to determine handling of the variables. Specifically,
valueType refers to the
\href{https://opaldoc.obiba.org/en/variables-data.html#value-types}{OBiBa data type of a variable}.
The valueType is specified in a data dictionary in a column 'valueType' and
can be associated with variables as attributes. Acceptable valueTypes
include 'text', 'integer', 'decimal', 'boolean', datetime', 'date'. The full
list of OBiBa valueType possibilities and their correspondence with R data
types are available using \link{valueType_list}.
}
\examples{
{

# use madshapR_DEMO provided by the package
library(dplyr)

###### Example 1: Combine functions and summarize datasets.
dossier <- list(iris = tibble())

dossier_summary <- dossier_summarize(dossier)
glimpse(dossier_summary)

}

}
