% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dictionaries_functions.R
\name{is_data_dict_shape}
\alias{is_data_dict_shape}
\title{Evaluate if any object is a potential data dictionary}
\usage{
is_data_dict_shape(object)
}
\arguments{
\item{object}{A potential data dictionary structure to be evaluated.}
}
\value{
A logical.
}
\description{
Confirms whether the input object is a valid enough or not to work with some
functions involving data dictionary shaping. This function mainly helps
validate input within other functions of the package but could be used to
check if a data dictionary minimal stricture is valid.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.
}
\examples{
{

# use DEMO_files provided by the package

data_dict <- DEMO_files$dd_MELBOURNE_1_format_maelstrom
is_data_dict_shape(data_dict)
is_data_dict_shape(iris)

}

}
\seealso{
For a better assessment, please use \code{\link[=data_dict_evaluate]{data_dict_evaluate()}}.
}
