% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataset_functions.R
\name{is_dataset}
\alias{is_dataset}
\title{Evaluate if any object is a dataset or not}
\usage{
is_dataset(object)
}
\arguments{
\item{object}{A potential dataset to be evaluated.}
}
\value{
A logical.
}
\description{
Confirms whether the input object is a valid dataset.
This function mainly helps validate input within other functions of the
package but could be used to check if a dataset is valid.
}
\details{
A dataset must be a data-frame or data-frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data dictionary
can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research standards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
{

# use DEMO_files provided by the package
# any data-frame (or tibble) can be a dataset by definition.

is_dataset(DEMO_files$dataset_MELBOURNE_1)
is_dataset(iris)
is_dataset(AirPassengers)

}

}
\seealso{
For a better assessment, please use \code{\link[=dataset_evaluate]{dataset_evaluate()}}.
}
