% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madshapR-package.R
\docType{package}
\name{madshapR-package}
\alias{madshapR}
\alias{madshapR-package}
\title{madshapR: Support Technical Processes Following 'Maelstrom Research' Standards}
\description{
Functions to support rigorous processes in data cleaning, evaluation, and documentation across datasets from different studies based on Maelstrom Research guidelines. The package includes the core functions to evaluate and format the main inputs that define the process, diagnose errors, and summarize and evaluate datasets and their associated data dictionaries. The main outputs are clean datasets and associated metadata, and tabular and visual summary reports. As described in Maelstrom Research guidelines for rigorous retrospective data harmonization (Fortier I and al. (2017) \doi{10.1093/ije/dyw075}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maelstrom-research/madshapR}
  \item Report bugs at \url{https://github.com/maelstrom-research/madshapR/issues}
}

}
\author{
\strong{Maintainer}: Guillaume Fabre \email{guijoseph.fabre@gmail.com} (\href{https://orcid.org/0000-0002-0124-9970}{ORCID})

Other contributors:
\itemize{
  \item Maelstrom-research group \email{info@maelstrom-research.org} [funder, copyright holder]
}

}
\keyword{internal}
