% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madshapR-package.R
\docType{package}
\name{madshapR-package}
\alias{madshapR}
\alias{madshapR-package}
\title{madshapR: Support Technical Processes Following 'Maelstrom Research' Standards}
\description{
The goal of this package is to help the user through the processes involving the Maelstrom Research guidelines. The user can: a) Create an R environment where the user can place files in an organized structure. b) Identify the location and content of files in user's environment. c) Generate bookdown reports of study-specific variables (input). e) Generate bookdown reports of study specific variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maelstrom-research/madshapR/ https://www.maelstrom-research.org}
  \item Report bugs at \url{https://github.com/maelstrom-research/madshapR/issues}
}

}
\author{
\strong{Maintainer}: Guillaume Fabre \email{gfabre@maelstrom-research.org} (\href{https://orcid.org/0000-0002-0124-9970}{ORCID})

Other contributors:
\itemize{
  \item Maelstrom-research group \email{info@maelstrom-research.org} [copyright holder, funder]
}

}
\keyword{internal}
