% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{madshapR_DEMO}
\alias{madshapR_DEMO}
\title{Built-in material allowing the user to test the package with demo data}
\format{
\subsection{\code{list}}{

A list with 12 elements used for testing the package
(data frames and lists):

\describe{
\item{data_dict_MELBOURNE}{Data dictionary of Melbourne
dataset}
\item{data_dict_PARIS}{Data dictionary of Paris dataset}
\item{data_dict_PARIS - collapsed}{Data dictionary of Paris with collapsed
categories}
\item{data_dict_TOKYO}{Data dictionary of Tokyo dataset}
\item{data_dict_TOKYO - errors}{Data dictionary of
Tokyo dataset containing errors for testing purpose}
\item{data_dict_TOKYO - errors with data}{data dictionary of Tokyo dataset
containing errors with Tokyo dataset for testing purpose}
\item{dataset_MELBOURNE}{Dataset of Melbourne}
\item{dataset_PARIS}{Dataset of Paris}
\item{dataset_TOKYO}{Dataset of Tokyo}
\item{dataset_TOKYO - errors with data}{dataset of Tokyo
containing errors with Tokyo data dictionary for testing purpose}
\item{taxonomy_PARIS}{Taxonomy of Paris dataset for testing purpose}
\item{dataset_suummary}{example of dataset summary for testing purpose}
...
}
}
}
\usage{
madshapR_DEMO
}
\description{
Built-in tibbles and lists allowing the user to test the package with demo
material.
}
\examples{
{

 print(madshapR_DEMO$dataset_TOKYO)

}
}
\keyword{datasets}
