% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{madshapR_DEMO}
\alias{madshapR_DEMO}
\title{Built-in material allowing the user to test the package with demo data}
\format{
\subsection{\code{list}}{

A list with 12 elements (data frames and lists) providing example objects
for testing the package:

\describe{
\item{data_dict_MELBOURNE}{Example Data dictionary for Melbourne dataset}
\item{data_dict_PARIS}{Example Data dictionary for Paris dataset}
\item{data_dict_PARIS - collapsed}{Example Data dictionary for Paris
with collapsed categories}
\item{data_dict_TOKYO}{Example Data dictionary for Tokyo dataset}
\item{data_dict_TOKYO - errors}{Data dictionary for Tokyo dataset
with errors}
\item{data_dict_TOKYO - errors with data}{Example Data Dictionary for Tokyo
dataset with errors with Tokyo dataset}
\item{dataset_MELBOURNE}{Example Dataset for MELBOURNE dataset}
\item{dataset_PARIS}{Example Dataset for PARIS dataset}
\item{dataset_TOKYO}{Example Dataset for TOKYO dataset}
\item{dataset_TOKYO - errors with data}{Example dataset of Tokyo
with errors with Tokyo data dictionary}
\item{taxonomy_PARIS}{Example Taxonomy for Paris dataset}
\item{dataset_summary}{Example of dataset summary}
...
}
}
}
\usage{
madshapR_DEMO
}
\description{
Demo datasets and data dictionaries, and taxonomy, to provide illustrative
examples of objects used by madshapR.
}
\examples{
{

 print(madshapR_DEMO$dataset_TOKYO)

}
}
\keyword{datasets}
