% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{madshapR_example}
\alias{madshapR_example}
\title{Built-in material allowing the user to test the package with example data}
\format{
\subsection{\code{list}}{

A list with 9 elements (data frames and lists) providing example objects
for testing the package:

\describe{
\item{dataset_example}{Example Dataset for example dataset}
\item{data_dict_example - as_data_dict}{Example Data dictionary for example
dataset where the structure is a data dictionary}
\item{data_dict_example - as_data_dict_mlstr}{Example Data dictionary for
example dataset where the structure is a data dictionary compliant with
Maelstrom}
\item{dataset_example - errors with data}{Example dataset of example
with errors with example data dictionary}
\item{data_dict_example - errors with data}{Example Data Dictionary for example
dataset with errors with example dataset}
\item{data_dict_example - errors}{Data dictionary for example dataset
with errors}
\item{data_dict_example - collapsed}{Example Data dictionary for example
with collapsed categories}
\item{taxonomy_example}{Example Taxonomy for example dataset}
\item{summary - dataset_example}{Example of dataset summary}
...
}
}
}
\usage{
madshapR_example
}
\description{
Example datasets and data dictionaries, and taxonomy, to provide illustrative
examples of objects used by madshapR.
}
\examples{
{
 
library(dplyr)
 
head(madshapR_example$`dataset_example`)
glimpse(madshapR_example$`data_dict_example`)

}
}
\keyword{datasets}
