% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-opal_functions.R
\name{opal_files_push}
\alias{opal_files_push}
\title{Upload files to an Opal environment}
\usage{
opal_files_push(opal, from, to)
}
\arguments{
\item{opal}{Opal login attributes.}

\item{from}{A character string of a path where the files will be taken from
in R.}

\item{to}{A character string of a path where the files will be placed to in
Opal.}
}
\value{
Folder(s) containing files coming from the user R environment in Opal.
The path to Opal needs to be pasted with Opal absolute path.
}
\description{
Uploads files from local to Opal environment. It is a wrapper of
\code{\link[opalr:opal.file_upload]{opalr::opal.file_upload()}}.
}
\details{
The user must be allowed to interact with their Opal. The errors
provided may be associated to the handler or the read/write
permissions to Opal. The user must have adequate credentials
to interact with their Opal environment.
}
\examples{
\dontrun{
# use case 1: place all files in a project (""home/project/"") or a user (""home/administrator/"")
opal_files_push(
opal = o,
from = ""DEMO"",
to = ""home/project/DEMO"")
}

}
\seealso{
Please see \href{https://opaldoc.obiba.org/en/dev/}{Opal documentation}for further
information.
}
