% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-opal_functions.R
\name{opal_tables_pull}
\alias{opal_tables_pull}
\title{Download tables from an Opal project as a study}
\usage{
opal_tables_pull(
  opal,
  project,
  table_list = NULL,
  content = c("dataset", "data_dict"),
  keep_as_study = TRUE,
  .remove_id = FALSE
)
}
\arguments{
\item{opal}{Opal login attributes.}

\item{project}{A character string specifying the Opal project name.}

\item{table_list}{A vector character string specifying Opal tables name.}

\item{content}{A vector of character string which indicates if the
function returns a dataset, or data dictionary. Default is 'dataset'.}

\item{keep_as_study}{whether to return a study or a dataset if there is only
one table. TRUE by default, if FALSE returns dataset.}
}
\value{
R objects (tibbles and list of tibbles) representing tables and their
respective data dictionary.
}
\description{
Downloads a study or dataset(s) from Opal project to local
environment. It is a wrapper of \code{\link[opalr:opal.table_get]{opalr::opal.table_get()}} and
\code{\link[opalr:opal.table_dictionary_get]{opalr::opal.table_dictionary_get()}}.
}
\details{
The user must be allowed to interact with their Opal. The errors
provided may be associated to the handler or the read/write
permissions to Opal. The user must have adequate credentials
to interact with their Opal environment.
}
\examples{
\dontrun{
# use case 1: download a table and its data dictionary associated
opal_tables_pull(opal = o,project = ""DEMO"", table = ""study_PARIS"")
# use case 2: download all tables and their data dictionaries associated
opal_tables_pull(opal = o,project = ""DEMO"")
}

}
\seealso{
Please see \href{https://opaldoc.obiba.org/en/dev/}{Opal documentation}for further
information.
}
