% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-opal_functions.R
\name{opal_tables_push}
\alias{opal_tables_push}
\title{Upload datasets into an Opal environment as tables in an Opal project}
\usage{
opal_tables_push(
  opal,
  dossier = NULL,
  data_dict = NULL,
  dataset = NULL,
  table_name = NULL,
  project_name,
  .force = FALSE,
  .overwrite = FALSE
)
}
\arguments{
\item{opal}{Opal login attributes.}

\item{dossier}{List of tibble, each of them being datasets.}

\item{data_dict}{A list of tibble(s) representing meta data of an
associated dataset. Automatically generated if not provided.}

\item{dataset}{A tibble identifying the input dataset observations associated to
its data dictionary.}

\item{table_name}{A character string specifying an Opal table name.}

\item{project_name}{A character string specifying the Opal project name.}

\item{.force}{If the destination already exists, stop with an informative
message if this flag is FALSE (default).}

\item{.overwrite}{If the destination table already exists, it will be
replaced (deleted, re-created with associated permissions reinstated and then
imported). Otherwise the table will be updated (data dictionaries merge may
conflict). Default is FALSE.}
}
\value{
A table in Opal.
}
\description{
Uploads a dossier or dataset(s) from local environment to Opal
environment. It is a wrapper of \code{\link[opalr:opal.table_create]{opalr::opal.table_create()}},
\code{\link[opalr:opal.table_save]{opalr::opal.table_save()}} and \code{\link[opalr:opal.table_dictionary_update]{opalr::opal.table_dictionary_update()}}.
}
\details{
The user must be allowed to interact with their Opal. The errors
provided may be associated to the handler or the read/write
permissions to Opal. The user must have adequate credentials
to interact with their Opal environment.
}
\examples{
{

#' # use DEMO_files provided by the package
library(opalr)
library(stringr)

dossier <- DEMO_files[stringr::str_detect(names(DEMO_files),"dataset_MELBOURNE")]

opal <- 
  opal.login('administrator','password', 
    url ='https://opal-demo.maelstrom-research.org')
    
tempdir <- basename(tempdir())
try(opal_project_create(opal, tempdir))

###### Example 1: push a table in a project.
try( 
  opal_tables_push(
  opal,
  dataset = dossier$dataset_MELBOURNE_1,
  table_name = 'dataset_MELBOURNE_1',
  project_name = tempdir,
  .force = TRUE,
  .overwrite = TRUE))
  
###### Example 2: push a dossier in a project.
try(
  opal_tables_push(
  opal, dossier, project_name = tempdir, .force = TRUE, .overwrite = TRUE))

}

}
\seealso{
Please see \href{https://opaldoc.obiba.org/en/dev/}{Opal documentation}for further
information.
}
