% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data_visualize.R
\name{open_visual_report}
\alias{open_visual_report}
\title{Open a visual report in a browser}
\usage{
open_visual_report(report_path)
}
\arguments{
\item{report_path}{A character string specifying the path of the report to
be opened.}
}
\description{
Opens a previously generated HTML bookdown document. This is a shortcut
function to access an existing visual report.
}
\examples{
{

# use DEMO_files provided by the package
library(dplyr)
library(fabR)  # silently_run

###### Example any data frame (or tibble) can be a dataset by definition.

to <- tempdir()
# Silencing the process, remove silently_run() to see the full process in 
# the console. 
silently_run(dataset_visualize(iris, to = to))

open_visual_report(to)

}

}
\seealso{
\code{\link[=dataset_visualize]{dataset_visualize()}}
}
