% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data_visualize.R
\name{open_visual_report}
\alias{open_visual_report}
\title{Open a visual report in a browser}
\usage{
open_visual_report(report_name)
}
\arguments{
\item{report_name}{A character string specifying the name of the report (a
folder in users environment) to be opened.}
}
\description{
The visual report previously generated in an HTML bookdown document can be
opened using this short-cut function.
This report can be used to assist the user in the assessment of the data
structure, fields investigation (mandatory or not), coherence across elements
and taxonomy or standard evaluation. The summaries and figures associated
provide dataset composition, with observation distribution and descriptive
statistics.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[=dataset_visualize]{dataset_visualize()}}
}
