% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-data_summarise.R
\name{resume_variables}
\alias{resume_variables}
\title{Generate a tibble resuming all variables in a dataset including its data dictionary}
\usage{
resume_variables(data, data_dict = NULL)
}
\arguments{
\item{data}{A tibble identifying the input data observations associated to
its data dictionary.}

\item{data_dict}{A list of tibble(s) representing meta data of an
associated dataset. Automatically generated if not provided.}
}
\value{
A tibble providing summary elements of a dataset, including its values and data
dictionary elements.
}
\description{
Generates a tibble that aggregates all columns
in a dataset with (if any) its data dictionary. The data dictionary (if present
separates observations between open values, missing values, categorical values ,
and categorical missing values (which corresponds to the 'missing' column in the
'Categories' sheet).
This internal function is used inside summary functions.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A dataset must be a data frame or data frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data
dictionary can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research stardards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[=summary_variables]{summary_variables()}}
}
