% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-opal_functions.R
\name{taxonomy_opal_mlstr_get}
\alias{taxonomy_opal_mlstr_get}
\title{Get available taxonomies in format used by Maelstrom Research}
\usage{
taxonomy_opal_mlstr_get(opal = NULL)
}
\arguments{
\item{opal}{Opal login attributes.}
}
\value{
A tibble identifying a taxonomy (generally generated from Opal taxonomy.
}
\description{
Downloads all taxonomies from an Opal server in a specific tibble format
used in the Maelstrom Research ecosystem.
}
\details{
Opal is the OBiBa application for data management, and Opal environments can
be integrated with RStudio environments using the package \code{opalr}.
The user must have adequate credentials to interact with a specific Opal
environment. Some function errors produced may be associated with the
handler or Opal read/write permissions.
}
\examples{
\dontrun{

library(opalr)
opal <- 
  opal.login('administrator','password', 
    url ='https://opal-demo.maelstrom-research.org/ui/index.html')
  
try(taxonomy_opal_mlstr_get(opal))
  
}

}
\seealso{
Please see \href{https://opaldoc.obiba.org/en/dev/}{Opal documentation}for further
information.
}
