% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-catalogue_functions.R
\name{valueType_adjust}
\alias{valueType_adjust}
\title{Replace the valueType from (or to) any data dictionary to (or from) dataset}
\usage{
valueType_adjust(from, to = NULL)
}
\arguments{
\item{from}{R object to be adjusted. Can be either a dataset or a data dictionary}

\item{to}{R object to be adjusted. Can be either a dataset or a data dictionary.
NULL by default, which is equivalent to valueType_self_adjust(... = from)}
}
\value{
Either a tibble, identifying the dataset, or a list of tibble(s)
identifying a data dictionary, depending which is 'to'.
}
\description{
Depending on the input provided (from and to can be either a dataset or a data
dictionary, this function takes the valueType of the first input (from), and assignes
column of a data dictionary of casted to a column in the dataset. If 'to' is not
provided, the function calls the function \code{datashapR::valueType_self_adjust(from)}
instead.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research stardards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A dataset must be a data frame or data frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data
dictionary can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research stardards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[=valueType_self_adjust]{valueType_self_adjust()}}
}
