% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-catalogue_functions.R
\name{valueType_guess}
\alias{valueType_guess}
\title{Guess the first possible valueType of any object}
\usage{
valueType_guess(x)
}
\arguments{
\item{x}{R object. Can be a vector.}
}
\value{
A character string which is the first possible valueType of the given
object.
}
\description{
Provides the first possible valueType, by trying to assign the object to
'boolean', then 'integer', then 'decimal', then 'date'.
If all fails, 'text', by default.
}
\details{
valueType is one of the property of a variable entity. It refers to
the (Obiba internal) type of any variable. The valueType can be 'text',
'integer', 'decimal', 'boolean', 'locale', 'datetime', 'date', 'binary',
'point', 'linestring', 'polygon'
The valueType list is available using \link{valueType_list}, and their
corresponding with typeof which is the (R internal) type of any object.
}
\examples{
\dontrun{
# use case 1: Apply valueType without specifying a data dictionary
as_valueType(""1"") \%>\% typeof()
tibble(iris \%>\% mutate(Species = as.character(Species))) \%>\%
mutate(across(c(starts_with(""Sepal"")), ~ as_valueType(.,""integer"")))
}

}
\seealso{
\href{https://opaldoc.obiba.org/en/dev/magma-user-guide/value/type.html}{Opal documentation}
}
