% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\docType{data}
\name{valueType_list}
\alias{valueType_list}
\title{Give list of authorized valueType values}
\format{
\subsection{\code{tibble}}{

A data-frame with 12 rows and 6 columns:
\describe{
\item{valueType}{data type as described in Opal}
\item{typeof}{data type provided by base::typeof}
\item{class}{data class provided by base::class}
\item{call}{function to transpose object according base::do.call function}
\item{toValueType}{ensemble data type as described in Opal}
\item{toTypeof}{ensemble data type provided by base::typeof}
\item{genericType}{ensemble data type which valueType belongs}
...
}
}
}
\usage{
valueType_list
}
\description{
Built-in tibbles showing the list of authorized valueType values, as
described in Opal, and their data type (R wise) corresponding. This tibble
is designed for programmatic use.
}
\details{
valueType is one of the property of a variable entity. It refers to
the (Obiba internal) type of any variable. The valueType can be 'text',
'integer', 'decimal', 'boolean', 'locale', 'datetime', 'date', 'binary',
'point', 'linestring', 'polygon'.
}
\examples{
{

valueType_list

}
}
\seealso{
\href{https://opaldoc.obiba.org/en/dev/magma-user-guide/value/type.html}{Opal documentation}
}
\keyword{datasets}
