% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\docType{data}
\name{valueType_list}
\alias{valueType_list}
\title{Built-in tibble of allowed valueType values}
\format{
\subsection{\code{tibble}}{

A data frame with 12 rows and 6 columns:
\describe{
\item{valueType}{data type as described in Opal}
\item{typeof}{data type provided by base::typeof}
\item{class}{data class provided by base::class}
\item{call}{function to transpose object according base::do.call function}
\item{toValueType}{ensemble data type as described in Opal}
\item{toTypeof}{ensemble data type provided by base::typeof}
\item{genericType}{ensemble data type which valueType belongs}
...
}
}
}
\usage{
valueType_list
}
\description{
Provides a built-in tibble showing the list of allowed Opal valueType values
and their corresponding R data types. This tibble is mainly used for
internal processes and programming.
}
\details{
The valueType is a property of a variable and is required in certain
functions to determine the handling of the variables. The valueType refers
to the OBiBa-internal type of a variable. It is specified in a data
dictionary in a column \code{valueType} and can be associated with variables as
attributes. Acceptable valueTypes include 'text', 'integer', 'decimal',
'boolean', datetime', 'date'). The full list of OBiBa valueType
possibilities and their correspondence with R data types are available using
\link{valueType_list}.
}
\examples{
{

valueType_list

}
}
\seealso{
\href{https://opaldoc.obiba.org/en/dev/magma-user-guide/value/type.html}{Opal documentation}
}
\keyword{datasets}
