% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-catalogue_functions.R
\name{valueType_self_adjust}
\alias{valueType_self_adjust}
\title{Search and replace the valueType of any data dictionary or dataset}
\usage{
valueType_self_adjust(...)
}
\arguments{
\item{...}{R object that can be either a dataset or a data dictionary.}
}
\value{
Either a tibble, identifying the dataset, or a list of tibble(s)
identifying a data dictionary, depending which the input refers to.
}
\description{
Using the function \code{\link[=valueType_guess]{valueType_guess()}}, this function provides the
first possible valueType, by trying to assign the object to 'boolean', then
'integer', then 'decimal', then 'date'. (If all fails, 'text', by default),
then either replace the valueType column of a data dictionary of casts a
column in the dataset. This depending on the input provided.
}
\details{
A data dictionary-like structure must be a list of at least one or two
data frame or data frame extension (e.g. a tibble) named 'Variables'
and 'Categories' (if any), representing meta data of an associated dataset.
The 'Variables' component must contain at least 'name' column and the
'Categories' component must at least contain 'variable' and 'name'
columns to be usable in any function of the package.
To be considered as a minimum (workable) data dictionary, it must also
have unique and non-null entries in 'name' column and the combination
'name'/'variable' must also be unique in 'Categories'.
In addition, the data dictionary may follow Maelstrom research standards,
and its content can be evaluated accordingly, such as naming convention
restriction, columns like 'valueType', 'missing' and 'label(:xx)',
and/or any taxonomy provided.

A dataset must be a data frame or data frame extension (e.g. a tibble) and
can be associated to a data dictionary. If not, a minimum workable data
dictionary can always be generated, when any column will be reported, and
any factor column will be analysed as categorical variable (the column
'levels' will be created for that. In addition, the dataset may follow
Maelstrom research standards, and its content can be evaluated accordingly,
such as naming convention restriction, or id columns declaration (which
full completeness is mandatory.
}
\examples{
\dontrun{
# Example 1: yyy yyy yyy.
}

}
\seealso{
\code{\link[=valueType_adjust]{valueType_adjust()}}
}
